Move type classes to its own file.
